separate library for olap modification schemas
