+++ ydb/core/tx/schemeshard/olap/layout/ya.make
@@ -1,0 +1,14 @@
+LIBRARY()
+
+SRCS(
+    layout.cpp
+)
+
+PEERDIR(
+    ydb/core/scheme
+    ydb/core/tablet_flat
+)
+
+YQL_LAST_ABI_VERSION()
+
+END()
