+++ ydb/core/tx/schemeshard/olap/manager/ya.make
@@ -1,0 +1,13 @@
+LIBRARY()
+
+SRCS(
+    manager.cpp
+)
+
+PEERDIR(
+    ydb/core/scheme
+    ydb/core/tx/schemeshard/olap/table
+    ydb/core/tx/schemeshard/olap/layout
+)
+
+END()
