+++ ydb/core/tx/schemeshard/olap/store/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+SRCS(
+    store.cpp
+)
+
+PEERDIR(
+    ydb/core/tx/schemeshard/olap/layout
+    ydb/core/protos
+)
+
+END()
