+++ ydb/core/tx/schemeshard/olap/table/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+SRCS(
+    table.cpp
+)
+
+PEERDIR(
+    ydb/core/protos
+    ydb/core/tablet_flat
+)
+
+END()
