+++ ydb/core/tx/schemeshard/olap/ya.make
@@ -8,6 +8,10 @@ PEERDIR(
     ydb/core/tx/schemeshard/olap/operations
     ydb/core/tx/schemeshard/olap/statistics
     ydb/core/tx/schemeshard/olap/options
+    ydb/core/tx/schemeshard/olap/layout
+    ydb/core/tx/schemeshard/olap/manager
+    ydb/core/tx/schemeshard/olap/store
+    ydb/core/tx/schemeshard/olap/table
 )
 
 END()
